requirements.txt
requests
selenium
opencv-python
Pillow
pytesseract